requirements.txt
luma==2.3.6
luma.core==2.4.1
luma.oled==3.13.0
Pillow==10.2.0
psutil==5.9.7
RPi.GPIO==0.7.1
